requirements.txt
APScheduler==3.10.4
Authlib==1.3.0
fastapi==0.111.0
python-dotenv==1.0.1
